stat

+++ ProjectAppjam/Assets/01. Scripts/Interface/IDamageable.cs
@@ -2,5 +2,5 @@ using UnityEngine;
 
 public interface IDamageable
 {
-    public void OnDamaged(int damage = 0, GameObject performer = null, Vector3 point = default);
+    public void OnDamaged(float damage = 0, GameObject performer = null, Vector3 point = default);
 }
